% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insidecpp11-package.R
\docType{package}
\name{insidecpp11-package}
\alias{insidecpp11}
\alias{insidecpp11-package}
\title{insidecpp11: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
