% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insidecpp11-package.R
\docType{package}
\name{insidecpp11-package}
\alias{insidecpp11}
\alias{insidecpp11-package}
\title{insidecpp11: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dis-organization/insidecpp11}
  \item Report bugs at \url{https://github.com/dis-organization/insidecpp11/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
